django/requirements.txt
Django==2.1.5
djangorestframework==3.9.0
django-cors-headers==2.4.0
psycopg2
djangorestframework-jwt
cryptography
python-jose
django-rest-knox
requests
mysqlclient
django-salesforce
jsonpickle
pandas
xlrd
openpyxl
django-rest-swagger
paramiko
pyjwt
pika
channels